ORM Mapping Many to one

+++ Spring-093-ORM-Mapping-ManyToOne/src/main/java/com/cybertek/repository/DepartmentRepository.java
@@ -4,6 +4,6 @@ import com.cybertek.entity.Department;
 import com.cybertek.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department,Long> {
+public interface DepartmentRepository extends JpaRepository<Department,Integer> {
     
 }
